+++ src/Playbook.Core/Leagues/LeagueType.cs
@@ -1,0 +1,8 @@
+namespace Playbook.Core.Leagues;
+
+public enum LeagueType
+{
+    Redraft = 0,
+    Dynasty = 1,
+    Keeper = 2
+}